Add rollup to build the library
Switch to using import and export instead of require() in src files.
Build files to lib/ and dist/.
Add index.js to combine the files into a single export.
Add babel config to package.json to transform modules when testing with jest, but not when building with rollup.
Remove webpack.
Switch to let and const in debug-harness.js.