Support arrays to specify a nested key path and score all keys if `keys` == []
Add sortKey option to specify the key name to sort on even if an empty keys array is passed in.
Rename defaultKeyName to sortKey.
Add tests for calling setKeys() after the constructor and scoring all keys on each object.
Add sortKey tests.
Add scoreValue string to result items.
Make it easier to get the highest matching value without having to traverse a path of sub-keys on the item to reach it.
Don't include a path on a key if the name is a single string wrapped in an array.
Add tests to check the new behavior.
Support dot and array paths in objects passed in the keys array.
Add a test for passing a dot path string in the name field in { name, scorer } objects passed in via the keys array.
Update JSDoc comments in QuickScore.js.
Update eslint to allow object properties on one line and multiline ternary operators.
Add tests for path arrays as keys and for ignoring dots in key names.
Update the documentation in QuickScore.js.
Use Array.isArray instead of instanceof.