Handle long, nearly-matching queries better
Add config.maxIterations to limit the number of loops that can be generated by long near matches.
Use the location param of indexOf to avoid creating a new string every time we look for the query.
Use default params in the Range constructor.
Add test of worst-case matching query that fits within maxIterations.
Add test of quickScore() with no params.